Fixes VSTS Bug 675488: [Feedback] Changing tab doesn't allow to directly type
https://devdiv.visualstudio.com/DevDiv/_workitems/edit/675488
