+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.Editor/TextEditorViewContent.cs
@@ -372,8 +372,10 @@ namespace MonoDevelop.Ide.Editor
 		}
 
 		#endregion
-	
 
-
+		public override void GrabFocus ()
+		{
+			textEditor.GrabFocus ();
+		}
 	}
 }
