+++ main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.Gui/SdiWorkspaceWindow.cs
@@ -262,9 +262,10 @@ namespace MonoDevelop.Ide.Gui
 		{
 			if (subViewToolbar != null)
 				subViewToolbar.Tabs [subViewToolbar.ActiveTab].Activate ();
+			SelectWindow ();
 		}
 		
-		public void SelectWindow()
+		public void SelectWindow ()
 		{
 			var window = tabControl.Toplevel as Gtk.Window;
 			if (window != null) {
@@ -287,11 +288,10 @@ namespace MonoDevelop.Ide.Gui
 			// The tab change must be done now to ensure that the content is created
 			// before exiting this method.
 			tabControl.CurrentTabIndex = tab.Index;
-
 			// Focus the tab in the next iteration since presenting the window may take some time
 			Application.Invoke ((o, args) => {
 				DockNotebook.ActiveNotebook = tabControl;
-				DeepGrabFocus (this.ActiveViewContent.Control);
+				ActiveViewContent.GrabFocus ();
 			});
 		}
 
@@ -341,60 +341,6 @@ namespace MonoDevelop.Ide.Gui
 			newTab.Content = this;
 			SetDockNotebook (nextNotebook, newTab);
 			SelectWindow ();
-		}
-		
-		static void DeepGrabFocus (Gtk.Widget widget)
-		{
-			Widget first = null;
-
-			foreach (var f in GetFocusableWidgets (widget)) {
-				if (f.HasFocus)
-					return;
-				
-				if (first == null)
-					first = f;
-			}
-			if (first != null) {
-				first.GrabFocus ();
-			}
-		}
-		
-		static IEnumerable<Gtk.Widget> GetFocusableWidgets (Gtk.Widget widget)
-		{
-			if (widget.CanFocus) {
-				yield return widget;
-			}
-
-			if (widget is Container c) {
-				foreach (var f in c.FocusChain.SelectMany (x => GetFocusableWidgets (x))) {
-					yield return f;
-				}
-
-				if (c.Children is var children) {
-					foreach (var f in children) {
-						if (f is Container container) {
-							foreach (var child in GetFocusableChildren (container)) {
-								yield return child;
-							}
-						}
-					}
-				}
-			}
-		}
-
-		static IEnumerable<Gtk.Widget> GetFocusableChildren (Gtk.Container container)
-		{
-			if (container.CanFocus) {
-				yield return container;
-			}
-
-			foreach (var f in container.Children) {
-				if (f is Container c) {
-					foreach (var child in GetFocusableChildren (c)) {
-						yield return child;
-					}
-				}
-			}
 		}
 
 		public DocumentToolbar GetToolbar (BaseViewContent targetView)
